scripts/python/requirements.txt
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-aiplatform>=1.71.0
pdfplumber==0.10.3
numpy>=2.0
python-dotenv==1.0.0
